CUDA backend: fix the type of int64_t.

+++ src/Futhark/CodeGen/ImpGen/Kernels/ToOpenCL.hs
@@ -295,7 +295,7 @@ genCUDAPrelude (OpenClRequirements _) =
 typedef char int8_t;
 typedef short int16_t;
 typedef int int32_t;
-typedef long int64_t;
+typedef long long int64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
